Filesystem.FAT: Rename Bytes_Total to File_Index
Bytes_Total feels like it holds the size of the file, as in "total bytes
of the file" but in really it's the current index in the file. And the
comment associated with this field already says that it's an index: "The
actual file index".

Offset of File_Offset are other possible names for this field.