Filesystem: Always use Status_Code as return value
as opposed to handles. And pass handles as out parameters.

File and directory handles will always be used for more than one
operation. Read/Write and Close at least, so it will require to declare
a variable for them.

   Handle : Any_File_Handle;
begin
   Open ("/my_file". Handle);
   Handle.Read (...)
   Handle.Close;

vs

  --  This is not practical because now I can't close the handle
  Open ("/my_file").Read (...);

On the other side, one might want to just check the status code of a
function without declaring a variable to do so.

if Open ("/my_file", Handle) /= OK then
   ...

vs

   Handle : Any_File_Handle;
   Status : Status_Code;
begin
   Handle := Open ("/my_file", Status);

   if Status /= OK then
      ...